pyproject.toml
[tool.poetry]
name = "balddetection"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
dvc = "^3.55.2"
tensorboard = "^2.18.0"
pytorch-lightning = "^2.4.0"
jupyter = "^1.1.1"
torch = {version = "^2.4.1+cu124", source = "pytorch"}
torchvision = {version = "^0.19.1+cu124", source = "pytorch"}
pyturbojpeg = "^1.7.7"
albumentations = "^1.4.18"
pandas = "^2.2.3"
seaborn = "^0.13.2"
numpy = "^2.1.2"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
